Add service and controller tests for be (#58)
* test(BE): add service and controller tests

* Maj(BE): add exceptions on all controllers EP

+++ backend/src/main/java/com/engineering/thesis/backend/serviceImpl/AppointmentServiceImpl.java
@@ -1,5 +1,6 @@
 package com.engineering.thesis.backend.serviceImpl;
 
+import com.engineering.thesis.backend.exception.ResourceNotFoundException;
 import com.engineering.thesis.backend.model.Appointment;
 import com.engineering.thesis.backend.repository.AppointmentRepository;
 import com.engineering.thesis.backend.service.AppointmentService;
@@ -7,6 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -14,13 +16,31 @@ public class AppointmentServiceImpl implements AppointmentService {
     private final AppointmentRepository appointmentRepository;
 
     @Override
-    public void create(Appointment appointment) {
-        appointmentRepository.save(appointment);
+    public Appointment create(Appointment appointment) throws ResourceNotFoundException {
+        Optional<Appointment> appointmentOptional = appointmentRepository.findById(appointment.getId());
+        if (appointmentOptional.isPresent()) {
+            throw new ResourceNotFoundException("Appointment with Id " + appointment.getId() + " already exists");
+        }
+        return appointmentRepository.save(appointment);
     }
 
     @Override
-    public void deleteById(Long id) {
+    public Appointment update(Appointment appointment) throws ResourceNotFoundException {
+        Optional<Appointment> priceOptional = appointmentRepository.findById(appointment.getId());
+        if (priceOptional.isEmpty()) {
+            throw new ResourceNotFoundException("Appointment with Id " + appointment.getId() + " doesn't  exists");
+        }
+        return appointmentRepository.save(appointment);
+    }
+
+    @Override
+    public Long deleteById(Long id) throws ResourceNotFoundException {
+        Optional<Appointment> priceOptional = appointmentRepository.findById(id);
+        if (priceOptional.isEmpty()) {
+            throw new ResourceNotFoundException("Appointment with Id " + id + " doesn't  exists");
+        }
         appointmentRepository.deleteById(id);
+        return id;
     }
 
     @Override
@@ -29,7 +49,11 @@ public class AppointmentServiceImpl implements AppointmentService {
     }
 
     @Override
-    public Appointment selectAppointmentById(Long id) {
-        return appointmentRepository.findById(id).get();
+    public Optional<Appointment> selectAppointmentById(Long id) throws ResourceNotFoundException {
+        Optional<Appointment> priceOptional = appointmentRepository.findById(id);
+        if (priceOptional.isEmpty()) {
+            throw new ResourceNotFoundException("Appointment with Id " + id + " doesn't  exists");
+        }
+        return appointmentRepository.findById(id);
     }
 }
