+++ backend/src/main/java/com/engineering/thesis/backend/serviceImpl/MedicalFacilityServiceImpl.java
@@ -1,5 +1,6 @@
 package com.engineering.thesis.backend.serviceImpl;
 
+import com.engineering.thesis.backend.exception.ResourceNotFoundException;
 import com.engineering.thesis.backend.model.MedicalFacility;
 import com.engineering.thesis.backend.repository.MedicalFacilityRepository;
 import com.engineering.thesis.backend.service.MedicalFacilityService;
@@ -7,6 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -14,13 +16,31 @@ public class MedicalFacilityServiceImpl implements MedicalFacilityService {
     private final MedicalFacilityRepository medicalFacilityRepository;
 
     @Override
-    public void create(MedicalFacility medicalFacility) {
-        medicalFacilityRepository.save(medicalFacility);
+    public MedicalFacility create(MedicalFacility medicalFacility) throws ResourceNotFoundException {
+        Optional<MedicalFacility> medicalFacilityOptional = medicalFacilityRepository.findById(medicalFacility.getId());
+        if (medicalFacilityOptional.isPresent()) {
+            throw new ResourceNotFoundException("Medical Facility with Id " + medicalFacility.getId() + " already exists");
+        }
+        return medicalFacilityRepository.save(medicalFacility);
     }
 
     @Override
-    public void deleteById(Long id) {
+    public MedicalFacility update(MedicalFacility medicalFacility) throws ResourceNotFoundException {
+        Optional<MedicalFacility> medicalFacilityOptional = medicalFacilityRepository.findById(medicalFacility.getId());
+        if (medicalFacilityOptional.isEmpty()) {
+            throw new ResourceNotFoundException("MedicalFacility with Id " + medicalFacility.getId() + " doesn't  exists");
+        }
+        return medicalFacilityRepository.save(medicalFacility);
+    }
+
+    @Override
+    public Long deleteById(Long id) throws ResourceNotFoundException {
+        Optional<MedicalFacility> medicalFacilityOptional = medicalFacilityRepository.findById(id);
+        if (medicalFacilityOptional.isEmpty()) {
+            throw new ResourceNotFoundException("MedicalFacility with Id " + id + " doesn't  exists");
+        }
         medicalFacilityRepository.deleteById(id);
+        return id;
     }
 
     @Override
@@ -29,7 +49,11 @@ public class MedicalFacilityServiceImpl implements MedicalFacilityService {
     }
 
     @Override
-    public MedicalFacility selectMedicalFacilityById(Long id) {
-        return medicalFacilityRepository.findById(id).get();
+    public Optional<MedicalFacility> selectMedicalFacilityById(Long id) throws ResourceNotFoundException {
+        Optional<MedicalFacility> medicalFacilityOptional = medicalFacilityRepository.findById(id);
+        if (medicalFacilityOptional.isEmpty()) {
+            throw new ResourceNotFoundException("MedicalFacility with Id " + id + " doesn't  exists");
+        }
+        return medicalFacilityRepository.findById(id);
     }
 }
