+++ backend/src/main/java/com/engineering/thesis/backend/serviceImpl/PatientMedicalDataServiceImpl.java
@@ -1,5 +1,6 @@
 package com.engineering.thesis.backend.serviceImpl;
 
+import com.engineering.thesis.backend.exception.ResourceNotFoundException;
 import com.engineering.thesis.backend.model.PatientMedicalData;
 import com.engineering.thesis.backend.repository.PatientMedicalDataRepository;
 import com.engineering.thesis.backend.service.PatientMedicalDataService;
@@ -7,6 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -14,13 +16,31 @@ public class PatientMedicalDataServiceImpl implements PatientMedicalDataService 
     private final PatientMedicalDataRepository patientMedicalDataRepository;
 
     @Override
-    public void create(PatientMedicalData patientMedicalData) {
-        patientMedicalDataRepository.save(patientMedicalData);
+    public PatientMedicalData create(PatientMedicalData patientMedicalData) throws ResourceNotFoundException {
+        Optional<PatientMedicalData> patientMedicalDataOptional = patientMedicalDataRepository.findById(patientMedicalData.getId());
+        if (patientMedicalDataOptional.isPresent()) {
+            throw new ResourceNotFoundException("PatientMedicalData with Id " + patientMedicalData.getId() + " already exists");
+        }
+        return patientMedicalDataRepository.save(patientMedicalData);
     }
 
     @Override
-    public void deleteById(Long id) {
+    public PatientMedicalData update(PatientMedicalData patientMedicalData) throws ResourceNotFoundException {
+        Optional<PatientMedicalData> patientMedicalDataOptional = patientMedicalDataRepository.findById(patientMedicalData.getId());
+        if (patientMedicalDataOptional.isEmpty()) {
+            throw new ResourceNotFoundException("PatientMedicalData with Id " + patientMedicalData.getId() + " doesn't  exists");
+        }
+        return patientMedicalDataRepository.save(patientMedicalData);
+    }
+
+    @Override
+    public Long deleteById(Long id) throws ResourceNotFoundException {
+        Optional<PatientMedicalData> patientMedicalDataOptional = patientMedicalDataRepository.findById(id);
+        if (patientMedicalDataOptional.isEmpty()) {
+            throw new ResourceNotFoundException("PatientMedicalData with Id " + id + " doesn't  exists");
+        }
         patientMedicalDataRepository.deleteById(id);
+        return id;
     }
 
     @Override
@@ -29,7 +49,11 @@ public class PatientMedicalDataServiceImpl implements PatientMedicalDataService 
     }
 
     @Override
-    public PatientMedicalData selectPatientMedicalDataById(Long id) {
-        return patientMedicalDataRepository.findById(id).get();
+    public Optional<PatientMedicalData> selectPatientMedicalDataById(Long id) throws ResourceNotFoundException {
+        Optional<PatientMedicalData> patientMedicalDataOptional = patientMedicalDataRepository.findById(id);
+        if (patientMedicalDataOptional.isEmpty()) {
+            throw new ResourceNotFoundException("PatientMedicalData with Id " + id + " doesn't  exists");
+        }
+        return patientMedicalDataRepository.findById(id);
     }
 }
