+++ backend/src/main/java/com/engineering/thesis/backend/serviceImpl/PriceServiceImpl.java
@@ -1,13 +1,14 @@
 package com.engineering.thesis.backend.serviceImpl;
 
+import com.engineering.thesis.backend.exception.ResourceNotFoundException;
 import com.engineering.thesis.backend.model.Price;
 import com.engineering.thesis.backend.repository.PriceRepository;
 import com.engineering.thesis.backend.service.PriceService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -15,13 +16,31 @@ public class PriceServiceImpl implements PriceService {
     private final PriceRepository priceRepository;
 
     @Override
-    public void create(Price price) {
-        priceRepository.save(price);
+    public Price create(Price price) throws ResourceNotFoundException {
+        Optional<Price> priceOptional = priceRepository.findById(price.getId());
+        if (priceOptional.isPresent()) {
+            throw new ResourceNotFoundException("Price with Id " + price.getId() + " already exists");
+        }
+        return priceRepository.save(price);
     }
 
     @Override
-    public void deleteById(Long id) {
+    public Price update(Price price) throws ResourceNotFoundException {
+        Optional<Price> priceOptional = priceRepository.findById(price.getId());
+        if (priceOptional.isEmpty()) {
+            throw new ResourceNotFoundException("Price with Id " + price.getId() + " doesn't  exists");
+        }
+        return priceRepository.save(price);
+    }
+
+    @Override
+    public Long deleteById(Long id) throws ResourceNotFoundException {
+        Optional<Price> priceOptional = priceRepository.findById(id);
+        if (priceOptional.isEmpty()) {
+            throw new ResourceNotFoundException("Price with Id " + id + " doesn't  exists");
+        }
         priceRepository.deleteById(id);
+        return id;
     }
 
     @Override
@@ -30,7 +49,11 @@ public class PriceServiceImpl implements PriceService {
     }
 
     @Override
-    public Price selectPriceById(Long id) {
-        return priceRepository.findById(id).get();
+    public Optional<Price> selectPriceById(Long id) throws ResourceNotFoundException {
+        Optional<Price> priceOptional = priceRepository.findById(id);
+        if (priceOptional.isEmpty()) {
+            throw new ResourceNotFoundException("Price with Id " + id + " doesn't  exists");
+        }
+        return priceRepository.findById(id);
     }
 }
